quiz_back/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.1
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
idna==3.11
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.6.2
